feat: run schema validation on create
