+++ main.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/zarf-dev/zarf/src/cmd"
 	"github.com/zarf-dev/zarf/src/config"
-	"github.com/zarf-dev/zarf/src/pkg/packager/lint"
+	"github.com/zarf-dev/zarf/src/pkg/lint"
 )
 
 //go:embed cosign.pub
