+++ src/pkg/packager/creator/normal.go
@@ -27,6 +27,7 @@ import (
 	"github.com/zarf-dev/zarf/src/internal/packager/kustomize"
 	"github.com/zarf-dev/zarf/src/internal/packager/sbom"
 	"github.com/zarf-dev/zarf/src/pkg/layout"
+	"github.com/zarf-dev/zarf/src/pkg/lint"
 	"github.com/zarf-dev/zarf/src/pkg/message"
 	"github.com/zarf-dev/zarf/src/pkg/packager/actions"
 	"github.com/zarf-dev/zarf/src/pkg/packager/filters"
@@ -74,7 +75,6 @@ func (pc *PackageCreator) LoadPackageDefinition(ctx context.Context, src *layout
 	if err != nil {
 		return types.ZarfPackage{}, nil, err
 	}
-
 	warnings = append(warnings, composeWarnings...)
 
 	// After components are composed, template the active package.
@@ -119,7 +119,8 @@ func (pc *PackageCreator) LoadPackageDefinition(ctx context.Context, src *layout
 		}
 	}
 
-	if err := pkg.Validate(); err != nil {
+	if findings, err := Validate(pkg); err != nil {
+		lint.PrintFindings(findings, lint.SevErr, pc.createOpts.BaseDir, pkg.Metadata.Name)
 		return types.ZarfPackage{}, nil, err
 	}
