+++ src/pkg/packager/creator/utils.go
@@ -5,14 +5,35 @@
 package creator
 
 import (
+	"fmt"
 	"os"
 	"runtime"
 	"time"
 
 	"github.com/zarf-dev/zarf/src/config"
+	"github.com/zarf-dev/zarf/src/pkg/lint"
 	"github.com/zarf-dev/zarf/src/pkg/packager/deprecated"
 	"github.com/zarf-dev/zarf/src/types"
 )
+
+// Validate errors if a package violates the schema or any runtime validations
+// This must be run while in the parent directory of the zarf.yaml being validated
+func Validate(pkg types.ZarfPackage) ([]lint.PackageFinding, error) {
+	if err := pkg.Validate(); err != nil {
+		return nil, fmt.Errorf("package validation failed: %w", err)
+	}
+
+	findings, err := lint.ValidateSchema()
+	if err != nil {
+		return nil, fmt.Errorf("unable to check schema: %w", err)
+	}
+
+	if lint.HasSevOrHigher(findings, lint.SevErr) {
+		return findings, fmt.Errorf("found errors in package")
+	}
+
+	return nil, nil
+}
 
 // recordPackageMetadata records various package metadata during package create.
 func recordPackageMetadata(pkg *types.ZarfPackage, createOpts types.ZarfCreateOptions) error {
